# services/orders/docs/swagger.yaml
basePath: /
definitions:
  data.Order:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      supplier_id:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
    type: object
  main.Response:
    properties:
      message:
        type: string
    type: object
host: localhost:8002
info:
  contact:
    email: omarbelghaouti@gmail.com
    name: Omar Belghaouti
  description: pdash orders service
  title: pdash orders service
  version: "1.0"
paths:
  /orders:
    get:
      consumes:
      - application/json
      description: Get all Orders
      operationId: get-orders
      parameters:
      - description: Supplier ID
        in: query
        name: supplier_id
        type: string
      - description: Customer ID
        in: query
        name: customer_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/data.Order'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response'
      summary: Get all Orders
    post:
      consumes:
      - application/json
      description: Create a new Order
      operationId: create-order
      parameters:
      - description: Order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/data.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/data.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response'
      summary: Create a new Order
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Order by ID
      operationId: delete-order-by-id
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response'
      summary: Delete a Order by ID
    get:
      consumes:
      - application/json
      description: Get a Order by ID
      operationId: get-order-by-id
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response'
      summary: Get a Order by ID
    put:
      consumes:
      - application/json
      description: Update a Order by ID
      operationId: update-order-by-id
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/data.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response'
      summary: Update a Order by ID
swagger: "2.0"